requirements.txt
# Core data science and ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Time-series forecasting
statsmodels>=0.13.0
prophet>=1.1.0

# Deep Learning
tensorflow>=2.7.0
keras>=2.7.0
torch>=1.10.0
pytorch-lightning>=1.5.0

# Jupyter and notebooks
jupyter>=1.0.0
jupyterlab>=3.0.0
ipython>=7.0.0
ipywidgets>=7.6.0

# API and web framework
Flask>=2.0.0
Flask-CORS>=3.0.0
Flask-RESTful>=0.3.9
Werkzeug>=2.0.0

# Task scheduling and orchestration
apache-airflow>=2.0.0
apache-airflow-providers-python>=3.0.0

# Database and ORM
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=5.4.0
click>=8.0.0
tqdm>=4.62.0
loguru>=0.5.3

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-asyncio>=0.18.0

# Code quality
black>=21.7.0
flake8>=3.9.0
mypy>=0.910
isort>=5.9.0
